+++ src/symbolicmodel.jl
