+++ test/runtests.jl
@@ -1,4 +1,5 @@
 include("test_controlsystem.jl")
 include("test_gridsubspace.jl")
-include("test_settransitions.jl")
-include("test_transitionmap.jl")
+include("test_symbolicmodel.jl")
+include("test_fromcontrolsystem.jl")
+include("test_controllerreach.jl")
