+++ test/test_symbolicmodel.jl
@@ -1,4 +1,4 @@
-include("../abstraction.jl")
+include("../src/abstraction.jl")
 
 module TestMain
